base/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.4
openvino==2024.4.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
jinja2==3.1.2
